Wrapped road emission factors by an object
This way we can add more information like the speed profile category, the year and the vehicle type. While this information is (indirectly) supplied by the client in the URL, this is more in line with how we're working for farms so far: always return the code that is used as well.
Since this isn't really a category (unique by 1 code), did not inherit from the Category object for this wrapping object. Could perhaps combine something to make a surrogate unique code, but that seems overkill at the moment.